Make crypt and crypt_gensalt use thread-local output buffers.
This change makes crypt and crypt_gensalt as thread-safe as
they can be without changing their interfaces.  Solaris
already made this change, and it’s being discussed by glibc
(with suggestion that it should be pushed upstream to the C
and POSIX standards committees):
https://sourceware.org/ml/libc-alpha/2018-10/msg00437.html

Portable programs should still use the r-variants, though,
because this is not a guaranteed feature that is portable
to other implementations of these functions. Also it doesn’t
help threads to not clobber their corresponding output buffer
on a second call from within the same thread.